Jupiter_NH3_CH4_Images_Arrays_P3 added
Jupiter_NH3_CH4_Images_Arrays_P3 was added. This single file, with multiple modules provides a working image array output suitable for submission to ALPO-Japan or PVOL.  It works right now *only* for 2021/09/05 UT. Next up is to modularize it and add a section for context images such as RGB and NUV. Then to generalize it to work on other dates, and of course future dates for the 2022 apparition.