Image Array for Single Methane Band, also for Jupiter & Saturn
This version is fully functional on the first eight 2022 Jupiter observations from 8/10 to 9/5. It produces only a SINGLE methane panel (continuum, in-band, ratio), which is not efficient when using additional channels, e.g., 620, 730, and 889 nm. It requires creating totally separate image arrays just to show the separate methane observations. It was modified to work on the Saturn observation of 9/5, but was not tested on the Saturn observation of 8/10. This makes way for a new version that will accommodate multiple CH4 observation bands in a single image array.