Another pre-AGU Commit
Lots of changes, mostly small. Includes better handling of latitude limits on GRS_NH3_Comparison_2022.py. It also includes a correction to averaging for the Fletcher 2016 data in plot_TEXES_Groups_P3.py. This accommodates the fact that the latitude grids vary from group to group and that group 4 only has 59 latitudes instead of 61.