2023 Paper Version 7 Code Update
Many changes and additions, including addition of scatter plots 2020 Jupiter retrievals, better plots of filters for SCT and VLT including the Jovian albedo. MOST IMPORTANTLY I aligned all the computations for k_eff and calibrations with the proper filter shapes specific to SCT and VLT. SCT filters now incorporate the overall system performance, but this is done explicitly in line. I NEED TO CHECK IF IT'S DONE EVERYWHERE NEEDED. ALSO, THERE'S TONS OF HARDCODING OF KEY PARAMETERS AND REARCHITECTNIG IS NEEDED NOT TO HAVE ISSUES ACCIDENTAL ERRORS IF ITEMS NEED UPDATING AND AREN'T UPDATED EVERYWHERE.