AAS-EPSC Poster Version 1.0 Update
This software version was used to create the plots in the version 1.0 poster for the October 2023 AAS-EPSC meeting. HOWEVER, the overall state of the code is mixed. Some sections of code have not been fully cleaned up or brought up to the desired standard for a 'real' release. In particular, handling of batch processing has inconsistent parsing of batch names.