Last with native L2 FITS files - Added planetmap functions
This is the last release with system 2 longitude native in L2 and L3 FITS map files. After this I'll be converting to the defacto standard of system 3 in all levels of FITS map files. That will require all map plotting software to absorb the ability to 'roll' maps to the desired display longitude.

Also, I installed planetmapper software and added emission and incidence angle (mu and mu_0) back planes to the L3 maps. I'll need to be moving that upstream to L2 FITS maps. I map convert the color RGB maps directly out of WINJupos to system 3 and add backplanes to those also.