L2 FITS in System 3 Longitude
Implemented L2 FITS in System 3 Longitude with new, consistent naming convention. Updated catalog for L2, which is now more consistent with L3 and will probably be able to consolidate the two into a single catalog. Am able to run custom map series as batches for analysis and studies. And added the beginning of a WINJupos study on 'dark formations'. Verified that profile software still works.