Intermediate Commit - Applying VLT calibration to SCT Obs
Main work here is in applying VLT calibrations to SCT observations. This required rerunning of the pipeline including L2, L3, and plots. It also required calibrating the MUSE spectra when convolved with the SCT filters. Finally, I'm in the midst of updating the profiles codes. And will probably need to figure out why the Teifel EWs no longer match the results I'm getting for L2 profiles.