Research Intermediate Commit - CALIBRATION PROBLEM?
Substantial work done since the ESS paper submission, particularly on profiles and continguous mapping. I also experimented with using only the 632nm channel for continuum on NH3. HOWEVER, SOMEWHERE IN THE PROCESS, THE OVERALL CALIBRATION OF FNH3 HAS SHIFTED AND NOW I GET VALUES ~10-15% HIGHER THAN IN THE PAPER. THIS COMMIT IS TO ENABLE COMPARES THAT MAY REVEAL WHY.