Refactored Map Related Codes
I flattened the codes for making and plotting patches and contours and updated the ROI_Script and the continguous maps codes to use the new flattened functions. I also updated the ROI code so it doesn't need BOTH the get_map_collections and ROI_collections to function since they were duplicative.