Image Segmentation Complete - fNH3, Plumes, NEDFs
Image segmentation can now be performed using thresholds for fNH3, Plumes, and NEDFs. The results can be overplotted, with labels, on maps. Tabular output files in CSV format are created with the properties of each identified feature (region or blob). I also fixed the mapping of times to lat-lon for contiguous maps.